+++ b4n/ui/widgets/command_palette.rs
@@ -303,8 +303,8 @@ impl Responsive for CommandPalette {
             return ResponseEvent::Handled;
         }
 
-        if event.is_out(MouseEventKind::RightClick, self.select().area())
-            || event.is_out(MouseEventKind::LeftClick, self.select().area())
+        if event.is_out(MouseEventKind::LeftClick, self.select().area())
+            || (event.is_out(MouseEventKind::RightClick, self.select().area()) && self.is_mouse_menu)
         {
             self.is_visible = false;
             return if self.is_mouse_menu {
@@ -324,7 +324,7 @@ impl Responsive for CommandPalette {
             return self.process_enter_key();
         }
 
-        if event.is_in(MouseEventKind::RightClick, self.select().filter_area()) {
+        if event.is_mouse(MouseEventKind::RightClick) && self.select().is_filter_visible() {
             return self.insert_from_clipboard();
         }